全改造段階表示スイッチを汎用的なラベル付きスイッチにする
